fix: Added a wrapper flex div to the MaxWidthWrapper component to resolve the issue with aligning contents of the defined max-width at the center on bigger screens. The wrapper flex div has justify-center utility class which help to achieve the desired alignment for the MaxWidthWrapper.
Please note that no additional width utility is needed when the MaxWidthWrapper is called on the top level return for a component, but in cases where there is another tag called before the Wrapper(semantic tags, example in NavBar and Footer), note that the padding defined by this wrapper will be applied within the wrapping semantic tag which will make the displayed width lesser than others where MaxWidthWrapper is at the top level. To control this, give the MaxWidthWrapper an overriding w-full i.e. just include 'w-full' in the utility className listed for MaxWidthWrapper.

In conclusion to avoid the different sizes due to the need for a semantic tag before calling MaxWidthWrapper in a component, you can just add w-full to className in every MaxWidthWrapper, it has no effect nor harm when there is no wrapping tag available!