feat: Added more assets to further build the remaining UI for this page, especially, the ExtendedArticles component. Also, added semantic tags to wrap the MaxWidthWrapper in the app's components used on homepage.
In the index.ts inside the constants folder, added comments that could be helpful when determining what stays in the file or which data will eventually go into context.

Finally, the homepage is almost completed, the outstanding is the component with videos. I encountered problems rendering mp4 file with next.js video tag. I also attempted next-video module, but ran into trouble. The problem is connected to turbopack config. I hope to find a solution to this soon.