feat: Connected the backend post endpoint for creating a new user to the sign-up page of the frontend.
Firstly, moved zod validation schemas from both the LogInForm and the SignUpForm components to a new file in zodValidations/auth. This is to modularize the code. Each validation schema is then imported into the respective form components.

Installed tanstack/react-query module which has functions for context creation for signup data and hook creation. These facilitate writing operations on the backend while calling backend service from the frontend.

Created context (QueryProvider) with QueryClient and QueryClientProvider methods from tanstack/react-query. The context was consumed by the main layout.tsx of the blog app.

Created createUser service. This is the client that calls the backend post endpoint for userRoute. This aims to create a new user and save their registration data on the MongoDB database.

Created an authentication hook called useRegisterUser which, with the help of tanstack/react-query's useMutation(), is used to call the createUser service. This saves the data of a new user to the appropriate collection in the connected mongoDB database.