fix: UI modifications and corrections to useDeleteBlog and useEditBlog hooks.
Updated the editBlog and the deleteBlog services to work as expected. The double arguments taken by editBlog function was passed as a single object holding the 2 arguments. The useEditBlog and the useDeleteBlog hooks were then updated to also reflect the changes to the service functions.

Removed unused icons/buttons from the Navbar and routed the discover button to discover page on-click.

Added an additional check for invalid token format to the middleware decoding token in session.ts.