fix: adjusted the UIs related with creating blog and editing blogs to improve their look and form validations.
In QuillEditBlogForm and QuillCreateBlogForm, replaced border-red for inadequate blogContent warning with toaster alert warning. Set the state that triggers blogContent word length check in QuillEditBlogForm within the component to prevent conflicts, and set the defaultValue for blog category to be the current one already submitted also in the QuillEditBlogForm component.

Modified the CSS properties of the background overlay for the EditModal component. The reduction of its z-index from 88 to 50 has made the blog category drop down menu visible on the edit modal.