fix: added more checks in categories, latest articles, authors sections and login page that render text/toaster alert which inform users as appropriate in cases where there are no blogs available or inability to send the backend call to the backend.
Also thought of triggering a state saved in context to make the login button inactive when the sections rendering data from the backend already shows server error. Presently, I have not figured out how to separate error on login for password-username mismatch from inability to connect to the server. I thought disabling the login button altogether when we are aware of server downtime on the landing page may be more efficient, but haven't figured it out. Any suggestions on how to get this done will be highly appreciated.