fix: updated backend logic for creating comment to return blog object of concerned blog with populated commenter info. Used this info to include the name of the commenter, the date the comment was written in the frontend's comments list.
Updated comments list based on updates received when new comments is added from the frontend. Doing this has removed the windows.reload() to get newly added comments.

Finally, reset the comments textarea value to empty string on successful submission of a new comment, and updated comment's zod validation to a maximum of 250 characters.