feat: implemented profile edit functionality for the security and the edit profile tabs of profile page.
Updated the editPasswordFormSchema and the editUserProfileFormSchema to make them more suitable for edit operation, created the functions that call the backend patch service for the updates which were called with mutateAsync in the respective components, and added a new context state to hold user's info set on clicking profile button from Navbar's dropdown. This data is for their preview and comparison before the changes.