fix: Profile page clean-up plus update of user data for the edit-profile tab.
Cleaned up console.logs from the tabs of the profile page.

Refactored the way user data is fetched and prepopulated in the edit form. Initially, the getAUser service was called in the Navbar component and stored in state on clicking profile from account's drop down list. A downside to this was that the edit page loses data on refresh and the page crashes. Therefore, the data is now being fetched within the edit-profile tab itself and the getAUser call was deleted from the Navbar component. The page no longer crashes.