feat: integrated backend services that allow comment authors to revisit and modify their comments.
Edit comment modal was created for the editing and submission of updated comments while the delete comment modal is for the deletion of comments.

Additional action types were added to the AppContext to control the opening and closing of these modals as well as facilitate movement of data necessary for these functionalities.